refactor(spice): split block-level helpers into block_application.rs
`apply_block_postprocessing`, `update_flat_storage_head`, and
`gc_memtrie_roots` move from `chunk_application.rs` to a new
`block_application.rs`. Keeps each file scoped to one persistence
level — chunk-level vs block-level — as the block-level surface grows.
