+++ chain/chain/src/spice/mod.rs
@@ -1,3 +1,4 @@
+pub mod block_application;
 pub mod chain;
 pub mod chunk_application;
 pub mod chunk_validation;
